test: refactored test run
