+++ e2e/scenarios/jira/index.ts
@@ -1,11 +1,25 @@
-export { addComment } from './addComment.spec';
-export { attachFile } from './attachFile.spec';
-export { createIssue } from './createIssue.spec';
-export { updateDescription } from './updateDescription.spec';
-export { updateIssueStatus } from './updateIssueStatus.spec';
-export { viewCommentWithImage } from './viewCommentWithImage.spec';
-export { renameIssue } from './renameIssue.spec';
-export { assigningFlow } from './assigningFlow.spec';
-export { checkImageInDescription } from './checkImageInDescription.spec';
-export { updateLabelsFlow } from './updateLabelsFlow.spec';
-export { authFlowJira } from './authFlow.spec';
+import { addComment } from './addComment.spec';
+import { assigningFlow } from './assigningFlow.spec';
+import { attachFile } from './attachFile.spec';
+import { authFlowJira } from './authFlow.spec';
+import { checkImageInDescription } from './checkImageInDescription.spec';
+import { createIssue } from './createIssue.spec';
+import { renameIssue } from './renameIssue.spec';
+import { updateDescription } from './updateDescription.spec';
+import { updateIssueStatus } from './updateIssueStatus.spec';
+import { updateLabelsFlow } from './updateLabelsFlow.spec';
+import { viewCommentWithImage } from './viewCommentWithImage.spec';
+
+export const jiraScenarios = [
+    { name: 'Authenticate with Jira', run: authFlowJira },
+    { name: 'Create issue', run: createIssue },
+    { name: 'Rename issue', run: renameIssue },
+    { name: 'Update issue description', run: updateDescription },
+    { name: 'Update issue status', run: updateIssueStatus },
+    { name: 'Add comment to issue', run: addComment },
+    { name: 'View comment with image in issue', run: viewCommentWithImage },
+    { name: 'Attach file to issue', run: attachFile },
+    { name: 'Assigning issue to myself', run: assigningFlow },
+    { name: 'Add and remove existing labels', run: updateLabelsFlow },
+    { name: 'Check image in description', run: checkImageInDescription },
+];
