+++ e2e/tests/jira/jiraDC.spec.ts
@@ -1,31 +1,6 @@
 import { test } from '@playwright/test';
 import { authenticateWithJiraDC } from 'e2e/helpers';
-import {
-    addComment,
-    assigningFlow,
-    attachFile,
-    authFlowJira,
-    createIssue,
-    renameIssue,
-    updateDescription,
-    updateIssueStatus,
-    updateLabelsFlow,
-    viewCommentWithImage,
-} from 'e2e/scenarios/jira';
-
-export const jiraScenarios = [
-    { name: 'Authenticate with Jira', run: authFlowJira },
-    { name: 'Create issue', run: createIssue },
-    { name: 'Rename issue', run: renameIssue },
-    { name: 'Update issue description', run: updateDescription },
-    { name: 'Update issue status', run: updateIssueStatus },
-    { name: 'Add comment to issue', run: addComment },
-    { name: 'View comment with image in issue', run: viewCommentWithImage },
-    { name: 'Attach file to issue', run: attachFile },
-    { name: 'Assigning issue to myself', run: assigningFlow },
-    { name: 'Add and remove existing labels', run: updateLabelsFlow },
-    // TODO: Add checkImageInDescription scenario after AXON-814 is fixed
-];
+import { jiraScenarios } from 'e2e/scenarios/jira';
 
 test.describe('Jira DC', () => {
     for (const scenario of jiraScenarios) {
